refactor(lyrics): extract three hooks from LyricsPanel
Extract useLyricsAutoScroll, useAudiencePlainTextPaging, and
useAirPlayPendingGuard into dedicated hook files to reduce
LyricsPanel.tsx from ~542 to ~369 lines and improve separation
of concerns.

No behavioral changes — all existing tests pass.
